Update on plotting time
Power 5- this is an attempt to plot power vs time, but has an erronouse
way of using the crawler
power6- thought to have been an improvment but has the same flaw in
implementation, that is that the crawler has not been reset/
power7 - now the implementation has been improved but resluts are still
odd.
timeAxis3 - expirimenting with filling time arrays.
timeAxis4 - correction of the use of the cralwer by combining smaller
tasks into a larger loop. This is implemented in power7.